+++ src/main/java/blackjack/domain/card/cardElement/Denomination.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.card.cardElement;
 
 import java.util.Arrays;
